attempt at stuttering improvement #48925: Use target time as the last frame time, rather than measuring the time at CalcFrameTime

+++ Descent3/game.cpp
@@ -116,7 +116,7 @@ gamemode Gamemode_info;
 bool InitGameScript();
 void CloseGameScript();
 
-float GetFPS ()
+double GetFPS ()
 {
 	if (Frametime == 0.0f) { Frametime = 0.1f; }
 	return 1.0 / Frametime;
